Rename `createAgents` to `agentKeeper`
This is a more general name, and in keeping with the theme.
